+++ packages/components/molecules/f-mega-modal/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.1.0
+------------------------------
+*February 2, 2022*
+
+### Changed
+- `f-button` version.
+
 
 v4.0.2
 ------------------------------