+++ packages/components/pages/f-checkout/CHANGELOG.md
@@ -3,6 +3,29 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v3.19.0
+------------------------------
+*February 3, 2022*
+
+### Changed
+- Handle `getAvailableFulfilmentTimes` issues
+
+
+v3.18.1
+------------------------------
+*February 2, 2022*
+
+### Changed
+- Use old notes field for order notes
+
+
+v3.18.0
+------------------------------
+*February 2, 2022*
+### Changed
+- Address service to handle AU/NZ addresses.
+
+
 v3.17.0
 ------------------------------
 *February 2, 2022*